requirements.txt
click
Flask
flask-cors
grist_api
gunicorn
imaplib2
pandas
pre-commit
pyjwt
python-dotenv
requests
ruff
xlsxwriter
